Update getOauthParameters.js
This change is to allow the use case where a body parser is already part of the router.  In the case of RESTify the stream is no longer readable by the time it gets to the ```getPostData``` function.  Therefore the already parsed post data is skipped.  Frameworks like Express and RESTify have body parsers that set the ```body``` property.  As a fallback the ```getPostData``` function can check this function and pass back the result as the second parameter in the callback.